+++ printer/formatter/src/main/java/br/dev/dig/logger/printer/formatter/styles/LoggerFormatStyleStackTrace.java
@@ -1,0 +1,23 @@
+package br.dev.dig.logger.printer.formatter.styles;
+
+import org.jetbrains.annotations.NotNull;
+import org.jetbrains.annotations.Nullable;
+
+import java.io.PrintWriter;
+import java.io.StringWriter;
+import java.time.LocalDateTime;
+
+import br.dev.dig.logger.printer.formatter.LoggerFormatter;
+
+public final class LoggerFormatStyleStackTrace implements LoggerFormatter.Style {
+    @Override
+    public @NotNull String print(@NotNull LocalDateTime start, int level, @Nullable String tag, @Nullable CharSequence message, @Nullable Throwable t) {
+        if (t == null) {
+            return "";
+        }
+        final StringWriter sw = new StringWriter();
+        final PrintWriter pw = new PrintWriter(sw);
+        t.printStackTrace(pw);
+        return sw.toString();
+    }
+}