+++ printer/formatter/src/test/java/br/dev/dig/logger/printer/formatter/styles/LoggerFormatStyleTagTest.java
@@ -1,4 +1,4 @@
-package br.dev.dig.logger.printer.println.styles;
+package br.dev.dig.logger.printer.formatter.styles;
 
 import org.jetbrains.annotations.NotNull;
 import org.junit.jupiter.api.Assertions;
@@ -8,12 +8,12 @@ import java.time.LocalDateTime;
 
 import br.dev.dig.logger.Logger;
 
-class PrintlnStyleTagTest extends PrintlnStyleCommon<PrintlnStyleTag> {
+class LoggerFormatStyleTagTest extends LoggerFormatStyleCommon<LoggerFormatStyleTag> {
 
     @Override
     @NotNull
-    PrintlnStyleTag create() {
-        return new PrintlnStyleTag();
+    LoggerFormatStyleTag create() {
+        return new LoggerFormatStyleTag();
     }
 
     @Override
@@ -23,49 +23,49 @@ class PrintlnStyleTagTest extends PrintlnStyleCommon<PrintlnStyleTag> {
 
     @Test
     void log_Invalid() {
-        final String output = style.print(Logger.LEVEL_NONE, randString(), LocalDateTime.now(), randString(), randThrowable());
+        final String output = style.print(LocalDateTime.now(), Logger.LEVEL_NONE, randString(), randString(), randThrowable());
         Assertions.assertNotNull(output);
         Assertions.assertEquals("", output);
     }
 
     @Test
     void log_Verbose() {
-        final String output = style.print(Logger.LEVEL_VERBOSE, randString(), LocalDateTime.now(), randString(), randThrowable());
+        final String output = style.print(LocalDateTime.now(), Logger.LEVEL_VERBOSE, randString(), randString(), randThrowable());
         Assertions.assertNotNull(output);
         Assertions.assertEquals("VERBOSE", output);
     }
 
     @Test
     void log_Debug() {
-        final String output = style.print(Logger.LEVEL_DEBUG, randString(), LocalDateTime.now(), randString(), randThrowable());
+        final String output = style.print(LocalDateTime.now(), Logger.LEVEL_DEBUG, randString(), randString(), randThrowable());
         Assertions.assertNotNull(output);
         Assertions.assertEquals("DEBUG", output);
     }
 
     @Test
     void log_Info() {
-        final String output = style.print(Logger.LEVEL_INFO, randString(), LocalDateTime.now(), randString(), randThrowable());
+        final String output = style.print(LocalDateTime.now(), Logger.LEVEL_INFO, randString(), randString(), randThrowable());
         Assertions.assertNotNull(output);
         Assertions.assertEquals("INFO", output);
     }
 
     @Test
     void log_Warning() {
-        final String output = style.print(Logger.LEVEL_WARNING, randString(), LocalDateTime.now(), randString(), randThrowable());
+        final String output = style.print(LocalDateTime.now(), Logger.LEVEL_WARNING, randString(), randString(), randThrowable());
         Assertions.assertNotNull(output);
         Assertions.assertEquals("WARNING", output);
     }
 
     @Test
     void log_Error() {
-        final String output = style.print(Logger.LEVEL_ERROR, randString(), LocalDateTime.now(), randString(), randThrowable());
+        final String output = style.print(LocalDateTime.now(), Logger.LEVEL_ERROR, randString(), randString(), randThrowable());
         Assertions.assertNotNull(output);
         Assertions.assertEquals("ERROR", output);
     }
 
     @Test
     void log_Wtf() {
-        final String output = style.print(Logger.LEVEL_ASSERT, randString(), LocalDateTime.now(), randString(), randThrowable());
+        final String output = style.print(LocalDateTime.now(), Logger.LEVEL_ASSERT, randString(), randString(), randThrowable());
         Assertions.assertNotNull(output);
         Assertions.assertEquals("ASSERT", output);
     }