+++ printer/println/src/main/java/br/dev/dig/logger/printer/println/PrintlnFormatter.java
@@ -1,104 +1,13 @@
 package br.dev.dig.logger.printer.println;
 
 import org.jetbrains.annotations.NotNull;
-import org.jetbrains.annotations.Nullable;
 
-import java.security.InvalidParameterException;
-import java.time.LocalDateTime;
-import java.util.ArrayList;
-import java.util.Arrays;
-import java.util.LinkedList;
-import java.util.regex.Matcher;
-import java.util.regex.Pattern;
+import br.dev.dig.logger.printer.formatter.LoggerFormatter;
 
-import br.dev.dig.logger.intrinsics.Intrinsics;
-import br.dev.dig.logger.printer.println.styles.PrintlnStyleConstant;
-import br.dev.dig.logger.printer.println.styles.PrintlnStyleCurrentDate;
-import br.dev.dig.logger.printer.println.styles.PrintlnStyleCurrentTime;
-import br.dev.dig.logger.printer.println.styles.PrintlnStyleElapsedTime;
-import br.dev.dig.logger.printer.println.styles.PrintlnStyleMessage;
-import br.dev.dig.logger.printer.println.styles.PrintlnStyleStackTrace;
-import br.dev.dig.logger.printer.println.styles.PrintlnStyleTag;
-import br.dev.dig.logger.printer.println.styles.PrintlnStyleTagShort;
-import br.dev.dig.logger.printer.println.styles.PrintlnStyleThrowableMessage;
-
-public class PrintlnFormatter {
-
-    public interface Style {
-        @NotNull
-        String print(int level, @Nullable String tag, @NotNull final LocalDateTime start, @Nullable final CharSequence message, @Nullable final Throwable t);
-    }
-
-    @NotNull
-    public Iterable<Style> styles;
-
+@Deprecated
+// Use LoggerFormatter
+public class PrintlnFormatter extends LoggerFormatter {
     public PrintlnFormatter(@NotNull Iterable<Style> styles) {
-        this.styles = Intrinsics.parameterNotNull(styles, "Styles list must not be null");
+        super(styles);
     }
-
-    public static PrintlnFormatter simple() {
-        return new PrintlnFormatter(Arrays.asList(
-            new PrintlnStyleElapsedTime(),
-            new PrintlnStyleConstant(" ["),
-            new PrintlnStyleTagShort(),
-            new PrintlnStyleConstant("] "),
-            new PrintlnStyleMessage(),
-            new PrintlnStyleThrowableMessage()
-        ));
-    }
-
-    public static PrintlnFormatter parse(@NotNull final String format) {
-        Intrinsics.parameterNotNull(format,"Format must not be null");
-        final LinkedList<Style> styles = new LinkedList<>();
-        final Pattern pattern = Pattern.compile("(\\$\\{[a-z\\-]+})");
-        final Matcher matches = pattern.matcher(format);
-        int lastEnd = 0;
-        while (matches.find()) {
-            if (lastEnd < matches.start()) {
-                styles.add(new PrintlnStyleConstant(format.substring(lastEnd, matches.start())));
-            }
-            lastEnd = matches.end();
-            switch (matches.group()) {
-                case "${tag}":
-                    styles.add(new PrintlnStyleTag());
-                    break;
-                case "${tag-short}":
-                    styles.add(new PrintlnStyleTagShort());
-                    break;
-                case "${date}":
-                    styles.add(new PrintlnStyleCurrentDate());
-                    break;
-                case "${time}":
-                    styles.add(new PrintlnStyleCurrentTime());
-                    break;
-                case "${elapsed}":
-                    styles.add(new PrintlnStyleElapsedTime());
-                    break;
-                case "${message}":
-                    styles.add(new PrintlnStyleMessage());
-                    break;
-                case "${throwable}":
-                    styles.add(new PrintlnStyleThrowableMessage());
-                    break;
-                case "${stacktrace}":
-                    styles.add(new PrintlnStyleStackTrace());
-                    break;
-            }
-        }
-        if (lastEnd < format.length()) {
-            styles.add(new PrintlnStyleConstant(format.substring(lastEnd)));
-        }
-        return new PrintlnFormatter(new ArrayList<>(styles));
-    }
-
-    @NotNull
-    public CharSequence format(@NotNull final LocalDateTime start, int level, @Nullable final String tag, @Nullable final CharSequence message, @Nullable final Throwable throwable) {
-        Intrinsics.parameterNotNull(start, "Start LocalDateTime must not be null");
-        final StringBuilder builder = new StringBuilder();
-        for (final PrintlnFormatter.Style style : styles) {
-            builder.append(style.print(level, tag, start, message, throwable));
-        }
-        return builder;
-    }
-
 }