+++ printer/println/src/main/java/br/dev/dig/logger/printer/println/PrintlnLogger.java
@@ -5,32 +5,52 @@ import org.jetbrains.annotations.Nullable;
 import org.jetbrains.annotations.VisibleForTesting;
 
 import java.time.LocalDateTime;
+import java.util.Arrays;
 
 import br.dev.dig.logger.BaseLogger;
 import br.dev.dig.logger.intrinsics.Intrinsics;
+import br.dev.dig.logger.printer.formatter.LoggerFormatter;
+import br.dev.dig.logger.printer.formatter.styles.LoggerFormatStyleConstant;
+import br.dev.dig.logger.printer.formatter.styles.LoggerFormatStyleElapsedTime;
+import br.dev.dig.logger.printer.formatter.styles.LoggerFormatStyleMessage;
+import br.dev.dig.logger.printer.formatter.styles.LoggerFormatStyleTagShort;
+import br.dev.dig.logger.printer.formatter.styles.LoggerFormatStyleThrowableMessage;
 
 
 public final class PrintlnLogger implements BaseLogger {
 
     @NotNull
     @VisibleForTesting
-    final PrintlnFormatter formatter;
+    LoggerFormatter formatter;
     @NotNull
     @VisibleForTesting
     final LocalDateTime start = LocalDateTime.now();
 
-    public PrintlnLogger(@NotNull final PrintlnFormatter formatter) {
+    public static abstract class Formatter {
+        public static LoggerFormatter simple() {
+            return new LoggerFormatter(Arrays.asList(
+                new LoggerFormatStyleElapsedTime(),
+                new LoggerFormatStyleConstant(" ["),
+                new LoggerFormatStyleTagShort(),
+                new LoggerFormatStyleConstant("] "),
+                new LoggerFormatStyleMessage(),
+                new LoggerFormatStyleThrowableMessage()
+            ));
+        }
+    }
+
+    public PrintlnLogger(@NotNull final LoggerFormatter formatter) {
         this.formatter = Intrinsics.parameterNotNull(formatter, "Formatter must not be null");
     }
 
     @SuppressWarnings("unused")
     public PrintlnLogger(@NotNull final String format) {
-        this(PrintlnFormatter.parse(format));
+        this(new LoggerFormatter.Builder().add(format).build());
     }
 
     @SuppressWarnings("unused")
     public PrintlnLogger() {
-        this(PrintlnFormatter.simple());
+        this(Formatter.simple());
     }
 
     @Override