+++ printer/println/src/main/java/br/dev/dig/logger/printer/println/styles/PrintlnStyleCurrentDate.java
@@ -11,9 +11,11 @@ import java.time.format.DateTimeFormatter;
 import br.dev.dig.logger.intrinsics.Intrinsics;
 import br.dev.dig.logger.printer.println.PrintlnFormatter;
 
+@Deprecated
+// Use LoggerFormatter
 public final class PrintlnStyleCurrentDate implements PrintlnFormatter.Style {
     @Override
-    public @NotNull String print(int level, @Nullable String tag, @NotNull LocalDateTime start, @Nullable CharSequence message, @Nullable Throwable t) {
+    public @NotNull String print(@NotNull LocalDateTime start, int level, @Nullable String tag, @Nullable CharSequence message, @Nullable Throwable t) {
         return format(LocalDate.now());
     }